functions/requirements.txt
firebase-functions
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.0.0
firebase-admin>=6.2.0
google-cloud-firestore
stripe>=7.0.0
openai>=1.0.0
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
beautifulsoup4
mutagen
google-cloud-texttospeech
google-cloud-storage>=2.10.0
tenacity>=8.0.0
Pillow>=10.0.0
flask-compress>=1.13
imageio-ffmpeg
pydantic>=2.0.0
email-validator>=2.0.0
bcrypt>=4.0.0
cryptography>=41.0.0
